pyproject.toml
[tool.poetry]
name = "kot-dos-project"
version = "0.1.0"
description = "First project on python with poetry"
authors = ["Виталий Кот <lowkent2me@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0"
account = { git = "https://github.com/ggramal/dos14-gramovich-git-flow.git", tag="v1.0.0"}
flask = "^2.3.2"
Flask-SQLAlchemy = "^3.0"
psycopg2 = "^2.9"

[tool.poetry.group.dev]
optional = true
[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
